Revert "Chapter 7"
