+++ app/src/main/AndroidManifest.xml
@@ -5,9 +5,7 @@
     <!-- To access the internet -->
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-    <uses-permission android:name="android.permission.READ_SYNC_SETTINGS"/>
-    <uses-permission android:name="android.permission.WRITE_SYNC_SETTINGS"/>
-    <uses-permission android:name="android.permission.AUTHENTICATE_ACCOUNTS"/>
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -15,9 +13,7 @@
         android:theme="@style/AppTheme" >
         <!-- Register weather content provider -->
         <provider android:name=".data.WeatherProvider"
-            android:authorities="com.example.pbabu.sunshine.app"
-            android:exported="false"
-            android:syncable="true"/>
+            android:authorities="com.example.pbabu.sunshine.app"/>
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -45,29 +41,6 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.example.pbabu.sunshine.app.MainActivity" />
         </activity>
-
-        <!-- SyncAdapter's dummy authentication service -->
-        <service
-            android:name=".sync.SunshineAuthenticatorService">
-            <intent-filter>
-                <action android:name="android.accounts.AccountAuthenticator"/>
-            </intent-filter>
-            <meta-data
-                android:name="android.accounts.AccountAuthenticator"
-                android:resource="@xml/authenticator" />
-        </service>
-        <!-- The SyncAdapter service -->
-        <service
-            android:name=".sync.SunshineSyncService"
-            android:exported="true"
-            >
-            <intent-filter>
-                <action android:name="android.content.SyncAdapter" />
-            </intent-filter>
-            <meta-data
-                android:name="android.content.SyncAdapter"
-                android:resource="@xml/syncadapter" />
-        </service>
     </application>
 
 </manifest>
