+++ app/src/main/java/com/example/pbabu/sunshine/app/ForecastDetailActivity.java
@@ -1,7 +1,5 @@
 package com.example.pbabu.sunshine.app;
 
-import android.app.NotificationManager;
-import android.content.Context;
 import android.content.Intent;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
@@ -9,29 +7,21 @@ import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 
-import com.example.pbabu.sunshine.app.sync.SunshineSyncAdapter;
-
 
 public class ForecastDetailActivity extends ActionBarActivity {
 
     private static final String LOG_TAG = ForecastDetailActivity.class.getSimpleName();
-    private boolean mIsMetric;
-    public static final String DETAIL_FRAGMENT_TAG = "DFTAG";
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        //cancel if there is any notification with id:3004
-        NotificationManager notificationManager =
-                (NotificationManager)this.getSystemService(Context.NOTIFICATION_SERVICE);
-        notificationManager.cancel(SunshineSyncAdapter.WEATHER_NOTIFICATION_ID);
         Log.d(LOG_TAG, "ForecastDetailActivity.onCreate");
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_forecast_detail);
-        mIsMetric = Utility.isMetric(this);
         if(savedInstanceState == null) {
             ForecastDetailActivityFragment df = ForecastDetailActivityFragment.newInstance(getIntent().getData());
             getSupportFragmentManager()
                     .beginTransaction()
-                    .add(R.id.weather_detail_container, df, DETAIL_FRAGMENT_TAG)
+                    .add(R.id.weather_detail_container, df)
                     .commit();
         }
     }
@@ -46,13 +36,6 @@ public class ForecastDetailActivity extends ActionBarActivity {
     protected void onResume() {
         Log.d(LOG_TAG, "ForecastDetailActivity.onResume");
         super.onResume();
-        final boolean currentMetric = Utility.isMetric(this);
-        ForecastDetailActivityFragment df = (ForecastDetailActivityFragment)getSupportFragmentManager()
-                .findFragmentByTag(DETAIL_FRAGMENT_TAG);
-        if(mIsMetric != currentMetric) {
-            df.onUnitsChanged();
-            mIsMetric = currentMetric;
-        }
     }
 
     @Override
