+++ app/src/main/java/com/example/pbabu/sunshine/app/ForecastDetailActivityFragment.java
@@ -156,9 +156,8 @@ public class ForecastDetailActivityFragment extends Fragment implements LoaderMa
     public void onActivityCreated(Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
         Bundle args = getArguments();
-        if(args != null && args.getParcelable(ARG_URI_STR) != null) {
-            mUri = args.getParcelable(ARG_URI_STR);
-
+        if(args != null && args.getString(ARG_URI_STR) != null) {
+            mUri = Uri.parse(args.getString(ARG_URI_STR));
         }
         getLoaderManager().initLoader(FORECAST_LOADER, null, this);
     }
@@ -261,14 +260,6 @@ public class ForecastDetailActivityFragment extends Fragment implements LoaderMa
     public void onLocationChanged(String location) {
         //get the date from the current uri
         final long currentDate = WeatherContract.WeatherEntry.getDateFromUri(mUri);
-        mUri = WeatherContract.WeatherEntry.buildWeatherLocationWithDate(location, currentDate);
-        getLoaderManager().restartLoader(FORECAST_LOADER, null, this);
-    }
-
-    public void onUnitsChanged() {
-        //get the date from the current uri
-        final long currentDate = WeatherContract.WeatherEntry.getDateFromUri(mUri);
-        String location = Utility.getPreferredLocation(getActivity());
         mUri = WeatherContract.WeatherEntry.buildWeatherLocationWithDate(location, currentDate);
         getLoaderManager().restartLoader(FORECAST_LOADER, null, this);
     }
