+++ app/src/main/java/com/example/pbabu/sunshine/app/MainActivity.java
@@ -1,7 +1,5 @@
 package com.example.pbabu.sunshine.app;
 
-import android.app.NotificationManager;
-import android.content.Context;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.net.Uri;
@@ -15,14 +13,13 @@ import android.view.View;
 import android.widget.Toast;
 
 import com.example.pbabu.sunshine.app.data.WeatherContract;
-import com.example.pbabu.sunshine.app.sync.SunshineSyncAdapter;
 
 
 public class MainActivity extends ActionBarActivity implements ForecastFragment.Callback {
     private static final String LOG_TAG = MainActivity.class.getSimpleName();
     private String mLocation;
-    private boolean mIsMetric;
     private boolean mTwoPane;
+    private static final String DETAIL_FRAGMENT_TAG = "DFTAG";
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         Log.d(LOG_TAG, "MainActivity.onCreate");
@@ -31,10 +28,6 @@ public class MainActivity extends ActionBarActivity implements ForecastFragment.
         //set default preferences
         PreferenceManager.setDefaultValues(this, R.xml.preferences, false);
         mLocation = Utility.getPreferredLocation(this);
-        mIsMetric = Utility.isMetric(this);
-        //clean up any notification with id:3004
-        NotificationManager notificationManager = (NotificationManager)this.getSystemService(Context.NOTIFICATION_SERVICE);
-        notificationManager.cancel(SunshineSyncAdapter.WEATHER_NOTIFICATION_ID);
         if(findViewById(R.id.weather_detail_container) != null){
             mTwoPane = true;
             if(savedInstanceState == null) {
@@ -42,16 +35,13 @@ public class MainActivity extends ActionBarActivity implements ForecastFragment.
                         .beginTransaction()
                         .replace(R.id.weather_detail_container,
                                 new ForecastDetailActivityFragment(),
-                                ForecastDetailActivity.DETAIL_FRAGMENT_TAG)
+                                DETAIL_FRAGMENT_TAG)
                         .commit();
             }
         }else {
             mTwoPane = false;
             getSupportActionBar().setElevation(0f);
         }
-
-        //initialize syncadapter
-        SunshineSyncAdapter.initialize(this);
     }
 
     @Override
@@ -65,28 +55,18 @@ public class MainActivity extends ActionBarActivity implements ForecastFragment.
         Log.d(LOG_TAG, "MainActivity.onResume");
         super.onResume();
         final String locationSetting = Utility.getPreferredLocation(this);
-        ForecastFragment ff = (ForecastFragment)getSupportFragmentManager().findFragmentById(R.id.fragment_forecast);
-        ForecastDetailActivityFragment df = (ForecastDetailActivityFragment)getSupportFragmentManager()
-                .findFragmentByTag(ForecastDetailActivity.DETAIL_FRAGMENT_TAG);
         if(!mLocation.equals(locationSetting)){
+            //do something
+            ForecastFragment ff = (ForecastFragment)getSupportFragmentManager().findFragmentById(R.id.fragment_forecast);
             if(ff != null) {
                 ff.onLocationChanged();
             }
+            ForecastDetailActivityFragment df = (ForecastDetailActivityFragment)getSupportFragmentManager()
+                    .findFragmentByTag(DETAIL_FRAGMENT_TAG);
             if(df != null) {
                 df.onLocationChanged(locationSetting);
             }
             mLocation = locationSetting;
-        }else {
-            final boolean currentMetric = Utility.isMetric(this);
-            if(mIsMetric != currentMetric) {
-                if(ff != null) {
-                    ff.onUnitsChanged();
-                }
-                if(df != null) {
-                    df.onUnitsChanged();
-                }
-                mIsMetric = currentMetric;
-            }
         }
     }
 
@@ -136,7 +116,7 @@ public class MainActivity extends ActionBarActivity implements ForecastFragment.
             ForecastDetailActivityFragment detailFragment = ForecastDetailActivityFragment
                     .newInstance(detailWeatherUri);
             getSupportFragmentManager().beginTransaction()
-                    .replace(R.id.weather_detail_container, detailFragment, ForecastDetailActivity.DETAIL_FRAGMENT_TAG)
+                    .replace(R.id.weather_detail_container, detailFragment, DETAIL_FRAGMENT_TAG)
                     .commit();
         }else {
             Intent detailIntent = new Intent(this, ForecastDetailActivity.class);
