+++ app/src/main/res/menu/forecastfragment.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
+<item android:id="@+id/action_refresh"
+    android:title="@string/action_refersh"
+    app:showAsAction="never"/>
 <item android:id="@+id/action_view_loc_on_map"
     android:title="@string/action_view_loc_on_map"
     app:showAsAction="never"/>
