+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,8 @@
 <resources xmlns:xliff="http://schemas.android.com/apk/res-auto">
     <string name="app_name">Sunshine</string>
+    <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
+    <string name="action_refersh" translatable="false">Refresh</string>
     <string name="title_activity_forecast_detail">Details</string>
     <string name="title_activity_settings">Settings</string>
     <string name="pref_location_title">Location</string>
@@ -12,11 +14,6 @@
     <string name="pref_units_default" translatable="false">metric</string>
     <string name="pref_units_imperial" translatable="false">imperial</string>
     <string name="pref_units_metric" translatable="false">metric</string>
-    <string name="pref_notification_title">Notification</string>
-    <string name="pref_notification_key" translatable="false">pref_notification</string>
-    <string name="pref_notification_summary_on" translatable="false">Weather notification is On</string>
-    <string name="pref_notification_summary_off" translatable="false">Weather notification is Off</string>
-    <string name="pref_enable_notifications_default" translatable="false">true</string>
     <string name="action_view_loc_on_map">View Location on Map</string>
     <string name="action_share">Share</string>
     <!-- Date label when displaying today's weather forecast [CHAR LIMIT=20] -->
@@ -46,15 +43,4 @@
 
     <!-- Humidity format CHAR LIMIT=25]-->
     <string name="format_humidity">Humidity: <xliff:g id="humidity">%1.0f</xliff:g> %%</string>
-    <!-- Account type (for sync-adapter authenticator) -->
-    <!-- SyncAdapter related -->
-    <string name="sync_account_type">sunshine.pbabu.com</string>
-    <string name="content_authority">com.example.pbabu.sunshine.app</string>
-
-    <!-- Notification Format -->
-    <string name="format_notification">
-    Forecast: <xliff:g id="condition">%1$s</xliff:g> High: <xliff:g id="high">%2$s</xliff:g> Low: <xliff:g id="low">%3$s</xliff:g>
-    </string>
-    <!-- Strings related to Notification preference -->
-    <string name="pref_last_notification">last_notification</string>
 </resources>
