--- app/src/main/res/xml/authenticator.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
-    android:accountType="@string/sync_account_type"
-    android:icon="@mipmap/ic_launcher"
-    android:label="@string/app_name"
-    android:smallIcon="@mipmap/ic_launcher" />
